Began reworking file format
Documented the format more, in the _internal header.
I'm changing the design to get rid of blocks.
The index will now be a separate area of the file.

Changed custom_endian to compat/endian.h, to act as
a stand-in for the real one if the include path can't
find the real one.